SplitView example: File name changes
Cnly changing capitalization of files and not the file name causes issue in OSX, since it is ignore capitalization. This updates the file name, so there are not conflicts.
