+++ components/split-view/__docs__/storybook-stories.jsx
@@ -5,7 +5,7 @@ import { SPLIT_VIEW } from '../../../utilities/constants';
 import Base from '../__examples__/base';
 import BaseMultiple from '../__examples__/base-multiple';
 import ExternalState from '../__examples__/external-state';
-import CustomItemList from '../__examples__/custom-item-list';
+import CustomItemList from '../__examples__/custom-items-list';
 
 storiesOf(SPLIT_VIEW, module)
 	.addDecorator((getStory) => (
