+++ components/split-view/__examples__/custom-items-list.jsx
