+++ components/split-view/component.json
@@ -1,32 +1,32 @@
 {
-    "component": "split-view",
-    "status": "prod",
-    "display-name": "Split View",
+	"component": "split-view",
+	"status": "prod",
+	"display-name": "Split View",
 	"classKey": "SplitView",
-    "last-accessibility-review": {
-        "date-iso-8601": "2018/01/18",
-        "commit-sha": "ad6b6c6523ee21cada11be5f7ea4d99abc530726"
-    },
-    "last-slds-markup-review": {
-        "date-iso-8601": "2018/01/18",
-        "commit-sha": "ad6b6c6523ee21cada11be5f7ea4d99abc530726"
-    },
-    "SLDS-component-path": "/components/split-view",
-    "dependencies": [
-        {
-            "component": "header",
+	"last-accessibility-review": {
+		"date-iso-8601": "2018/01/18",
+		"commit-sha": "ad6b6c6523ee21cada11be5f7ea4d99abc530726"
+	},
+	"last-slds-markup-review": {
+		"date-iso-8601": "2018/01/18",
+		"commit-sha": "ad6b6c6523ee21cada11be5f7ea4d99abc530726"
+	},
+	"SLDS-component-path": "/components/split-view",
+	"dependencies": [
+		{
+			"component": "header",
 			"classKey": "SplitViewHeader"
 		},
-        {
-            "component": "listbox",
+		{
+			"component": "listbox",
 			"classKey": "SplitViewListbox"
 		}
-    ],
+	],
 	"site-stories": [
 		"/__examples__/base.jsx",
 		"/__examples__/base-multiple.jsx",
 		"/__examples__/external-state.jsx",
-		"/__examples__/custom-item-list.jsx"
+		"/__examples__/custom-items-list.jsx"
 	],
-    "url-slug": "split-view"
+	"url-slug": "split-view"
 }
